requirements.txt
requests        
python-dotenv
easse@ git+https://github.com/feralvam/easse.git
openai